@@ -3,20 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "3.7.0",
-    "@material-ui/icons": "3.0.1",
+    "@material-ui/core": "^3.9.3",
+    "@material-ui/icons": "^3.0.2",
     "axios": "0.18.0",
     "classnames": "^2.2.6",
-    "hex-rgb": "^3.0.0",
+    "hex-rgb": "^4.1.0",
     "merge": ">=1.2.1",
     "moment": "^2.24.0",
-    "react": "16.6.3",
-    "react-dom": "16.6.3",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
     "react-hotkeys": "^1.1.4",
     "react-jss": "^8.6.1",
-    "react-redux": "6.0.0",
-    "react-router": "4.3.1",
-    "react-router-dom": "4.3.1",
+    "react-redux": "^7.0.2",
+    "react-router": "^5.0.0",
+    "react-router-dom": "^5.0.0",
     "react-scripts": "^2.1.5",
     "redux": "^4.0.1",
     "typeface-roboto": "0.0.54"
@@ -29,9 +29,9 @@
   "proxy": "http://localhost",
   "devDependencies": {
     "merge": ">=1.2.1",
-    "mocha": "5.2.0",
-    "react-test-renderer": "16.6.3",
-    "sinon": "7.2.2"
+    "mocha": "^6.1.3",
+    "react-test-renderer": "^16.8.6",
+    "sinon": "^7.3.1"
   },
   "browserslist": [
     ">0.2%",
